feat(component): replaced  ->  component
